+++ doc/source/changes.rst
@@ -7,6 +7,9 @@ Change Log
   server when using the :class:`tunic.api.VirtualEnvInstallation` class.
 * Add :doc:`usage` section to the documentation that explains how to use
   each part of the library at a higher level than just the :doc:`api` section.
+* Add :class:`tunic.api.LocalArtifactTransfer` class for transfering locally
+  built artifacts to a remote server and cleaning them up after deployment
+  has completed.
 
 0.3.0 - 2014-09-28
 ------------------